requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pypdf==5.1.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy==2.2.1
openai==1.58.1
httpx==0.28.1
python-dotenv==1.0.1
pydantic-settings==2.7.1
aiofiles==24.1.0
